refactor(plugins/state): use the new actions under /admin
